+++ node/metrics/metrics.ts
@@ -1,0 +1,31 @@
+import axios from 'axios'
+
+const ANALYTICS_URL = 'https://rc.vtex.com/api/analytics/schemaless-events'
+
+type CreateQuoteMetric = {
+  kind: 'create-quote-graphql-event'
+  description: 'Create Quotation Action - Graphql'
+}
+
+type SendMessageMetric = {
+  kind: 'send-message-graphql-event'
+  description: 'Send Message Action - Graphql'
+}
+
+type UseQuoteMetric = {
+  kind: 'use-quote-graphql-event'
+  description: 'Use Quotation Action - Graphql'
+}
+
+export type Metric = {
+  name: 'b2b-suite-buyerorg-data'
+  account: string
+} & (CreateQuoteMetric | SendMessageMetric | UseQuoteMetric)
+
+export const sendMetric = async (metric: Metric) => {
+  try {
+    await axios.post(ANALYTICS_URL, metric)
+  } catch (error) {
+    console.warn('Unable to log metrics', error)
+  }
+}